requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
paho-mqtt==1.6.1
httpx==0.27.0
faster-whisper
broadlink>=0.19.0
requests>=2.31.0
